pyproject.toml
[project]
name = "upstox-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol server for Upstox trading functionality"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.11.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[project.scripts]
mcp-server = "mcpserver.__main__:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
